requirements.txt
Flask==2.3.3
pdfplumber==0.11.0
google-generativeai==0.7.0
python-dotenv==1.0.1